MinWeightedSum: changed Integer to Long

+++ src/com/littlewhywhat/algorithms/schedule/MinWeightedSum.java
@@ -3,12 +3,12 @@ package com.littlewhywhat.algorithms.schedule;
 import com.littlewhywhat.algorithms.AbstractAlgorithm;
 
 public class MinWeightedSum<E extends Job> extends
-		AbstractAlgorithm<Void, Schedule<E>, Integer> {
+		AbstractAlgorithm<Void, Schedule<E>, Long> {
 
 	@Override
 	public void execute() {
 		final Schedule<E> schedule = getData();
-		int result = 0;
+		long result = 0;
 		while (!schedule.isEmpty()) {
 			E job = schedule.poll();
 			result += job.getWeight() * schedule.getTime();
